requirements.txt
streamlit
groq
neo4j
sentence-transformers
yfinance
python-dotenv
streamlit-agraph
pyvis
requests
beautifulsoup4
pandas
pytest
pytest-mock
pytest-cov
torchvision
